:truck: rename: #33 구조 정리를 위해 폴더 이동 및 경로 수정
