+++ src/components/story/detail/comment/CommentContent.tsx
@@ -4,10 +4,19 @@ import {
   inputBoxStyle,
   inputWrapperStyle,
 } from '@components/story/detail/comment.recipe';
-import { CommentItem } from '@components/story/useComment';
+import { CommentItem } from '@components/story/hooks/useComment';
 import { Button } from '@components/ui/common/buttons/Button';
 
 import { css } from '@root/styled-system/css';
+
+type Props = {
+  comment: CommentItem;
+  isEditing: boolean;
+  value: string;
+  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  onSubmit: () => void;
+  onCancel: () => void;
+};
 
 // 댓글 내용을 보여주거나 수정 폼을 렌더링하는 컴포넌트 역할 분리
 function CommentContent({
@@ -17,14 +26,7 @@ function CommentContent({
   onChange,
   onSubmit,
   onCancel,
-}: {
-  comment: CommentItem;
-  isEditing: boolean;
-  value: string;
-  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
-  onSubmit: () => void;
-  onCancel: () => void;
-}) {
+}: Props) {
   if (!isEditing)
     return (
       <p className={css({ textAlign: 'start', mb: 4 })}>{comment.title}</p>
